Refactored the TopicTableManager to use the newly created TableCell classes.
